dawah-translate/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yt-dlp>=2026.3.0
faster-whisper==1.1.0
anthropic>=0.40.0
pysrt==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
aiofiles==24.1.0
requests>=2.31.0
deno>=2.0.0
